move generate.py into plasmapy_sphinx.automodsumm
